AppKitIntrospectionView
