+++ Introspect/AppKitIntrospectionView.swift
@@ -1,3 +1,4 @@
+#if canImport(AppKit)
 import SwiftUI
 import AppKit
 
@@ -21,7 +22,7 @@ public class IntrospectionNSView: NSView {
 
 /// Introspection View that is injected into the UIKit hierarchy alongside the target view.
 /// After `updateNSView` is called, it calls `selector` to find the target view, then `customize` when the target view is found.
-public struct IntrospectionView<TargetViewType: NSView>: NSViewRepresentable {
+public struct AppKitIntrospectionView<TargetViewType: NSView>: NSViewRepresentable {
     
     /// Method that introspects the view hierarchy to find the target view.
     /// First argument is the introspection view itself, which is contained in a view host alongside the target view.
@@ -38,7 +39,7 @@ public struct IntrospectionView<TargetViewType: NSView>: NSViewRepresentable {
         self.customize = customize
     }
     
-    public func makeNSView(context: NSViewRepresentableContext<IntrospectionView>) -> IntrospectionNSView {
+    public func makeNSView(context: NSViewRepresentableContext<AppKitIntrospectionView>) -> IntrospectionNSView {
         let view = IntrospectionNSView()
         view.setAccessibilityLabel("IntrospectionNSView<\(TargetViewType.self)>")
         return view
@@ -51,7 +52,7 @@ public struct IntrospectionView<TargetViewType: NSView>: NSViewRepresentable {
     /// gets called when the introspection view gets removed from the hierarchy.
     public func updateNSView(
         _ nsView: IntrospectionNSView,
-        context: NSViewRepresentableContext<IntrospectionView>
+        context: NSViewRepresentableContext<AppKitIntrospectionView>
     ) {
         DispatchQueue.main.async {
             guard let targetView = self.selector(nsView) else {
@@ -61,4 +62,5 @@ public struct IntrospectionView<TargetViewType: NSView>: NSViewRepresentable {
         }
     }
 }
+#endif
